[update][m] base.html change inline js code
According to the client's requirements, I have changed the behavior of the Search box in relation to the use of the Enter command from the keyboard. Now, the first press of Enter selects Search datasets and the second executes a search execution.
I also changed the behavior of the Search box by pressing the Escape command on the keyboard. Earlier in the search box Escape was printed in the extension of the search term, and now the first press of Esc returns the focus to the search box, and the second closes the search box list. Similar to Cancel command.