[update][s] t2 - preserve spaces and new lines in descriptions (#13)
* [update][s] preserve spaces and new lines in descriptions

Spaces and new lines are not showing in dataset descriptions.

* [update][s] moving keywords section below metadata

Client wanted that Keyword section to be moved below Metadata section

* [update][s] group information within metadata

Client wanted that Group information should be visible within metadata.

* [update][s] click on resourcename sends to preview

Client wants that click on Resource name begins download instead sends to Preview.

* [update][s] T14 - no preview button

Client wants that `Preview` button not showing when there is no views for particular dataset.

* [del][xs] deleted showcase.html from root

* [upload][xs] upload showcase.html at right place - views folder

* [update][xs] coloring filetypes buttons

I added coloring for two file types and corrected not coloring for `GeoJSON`

* [update][xs] increased paddings in moved keyword section

I increased padding and size of keyword badges in Keyword section.

* [update][xs] groups to be clickable

Now Groups items are clickable. Click on item create query with that search term on all datasets.

* [update][xs] ODDK frontend-v2 v1.0.0 provides downloadPath

You are right. This was long time ago, but if I remember well, I tested this in other browsers and noticed that download links not forming well in IE with `downloadPath` parameter. Anyway, I revert this changes as you suggested.

* [update][xs] T17 - the top breadcrumb does not work

I corrected top breadcrumb on `showcase.html` page not working. Instead of `/search/dataset-title` I placed `/organization-name/dataset-title`

* [xs] fixed error on commit

I tried to see what is an impact of using `dataset` instead `resource` and that is a cause of this mistake.

* [fix][xs] `/collections/{{group.name} instead of query

I fixed this as you suggested.

* [fix] using `owner.name` in breadcrumb

I made fix and changed `dataset.organization.name` to ` owner.name`.